backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
click==8.1.7
comm==0.2.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-datatables==0.7.2
executing==2.0.1
fonttools==4.37.4
git-filter-repo==2.45.0
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
jiggler==0.0.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
networkx==3.4.2
npm==0.1.1
numpy==1.26.0
optional-django==0.1.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.9.0
Pillow==9.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.4
pyzmq==25.1.2
scipy==1.13.0
seaborn==0.12.0
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tornado==6.4
tqdm==4.67.0
traitlets==5.14.1
wcwidth==0.2.13
